Fix TypeScript strict mode errors and add tsconfig

+++ demo/main.ts
@@ -8,32 +8,43 @@ namespace appn.demo {
         private tempblock2: HTMLElement | null = null;
         private aclick: boolean = false;
         private noinfo: boolean = false;
-        private flowyInstance: appn.flowy.Flowy;
+        private flowyInstance!: appn.flowy.Flowy;
 
         constructor() {
             this.init();
         }
 
         private init(): void {
-            document.getElementById("blocklist").innerHTML = '<div class="blockelem create-flowy noselect"><input type="hidden" name="blockelemtype" class="blockelemtype" value="1"><div class="grabme"><img src="assets/grabme.svg"></div><div class="blockin">                  <div class="blockico"><span></span><img src="assets/eye.svg"></div><div class="blocktext">                        <p class="blocktitle">New visitor</p><p class="blockdesc">Triggers when somebody visits a specified page</p>        </div></div></div><div class="blockelem create-flowy noselect"><input type="hidden" name="blockelemtype" class="blockelemtype" value="2"><div class="grabme"><img src="assets/grabme.svg"></div><div class="blockin">                    <div class="blockico"><span></span><img src="assets/action.svg"></div><div class="blocktext">                        <p class="blocktitle">Action is performed</p><p class="blockdesc">Triggers when somebody performs a specified action</p></div></div></div><div class="blockelem create-flowy noselect"><input type="hidden" name="blockelemtype" class="blockelemtype" value="3"><div class="grabme"><img src="assets/grabme.svg"></div><div class="blockin">                    <div class="blockico"><span></span><img src="assets/time.svg"></div><div class="blocktext">                        <p class="blocktitle">Time has passed</p><p class="blockdesc">Triggers after a specified amount of time</p>          </div></div></div><div class="blockelem create-flowy noselect"><input type="hidden" name="blockelemtype" class="blockelemtype" value="4"><div class="grabme"><img src="assets/grabme.svg"></div><div class="blockin">                    <div class="blockico"><span></span><img src="assets/error.svg"></div><div class="blocktext">                        <p class="blocktitle">Error prompt</p><p class="blockdesc">Triggers when a specified error happens</p>              </div></div></div>';
-
-            this.flowyInstance = appn.flowy.create(document.getElementById("canvas"), this.drag.bind(this), this.release.bind(this), this.snapping.bind(this));
+            const blocklist = document.getElementById("blocklist");
+            if (blocklist) {
+                blocklist.innerHTML = '<div class="blockelem create-flowy noselect"><input type="hidden" name="blockelemtype" class="blockelemtype" value="1"><div class="grabme"><img src="assets/grabme.svg"></div><div class="blockin">                  <div class="blockico"><span></span><img src="assets/eye.svg"></div><div class="blocktext">                        <p class="blocktitle">New visitor</p><p class="blockdesc">Triggers when somebody visits a specified page</p>        </div></div></div><div class="blockelem create-flowy noselect"><input type="hidden" name="blockelemtype" class="blockelemtype" value="2"><div class="grabme"><img src="assets/grabme.svg"></div><div class="blockin">                    <div class="blockico"><span></span><img src="assets/action.svg"></div><div class="blocktext">                        <p class="blocktitle">Action is performed</p><p class="blockdesc">Triggers when somebody performs a specified action</p></div></div></div><div class="blockelem create-flowy noselect"><input type="hidden" name="blockelemtype" class="blockelemtype" value="3"><div class="grabme"><img src="assets/grabme.svg"></div><div class="blockin">                    <div class="blockico"><span></span><img src="assets/time.svg"></div><div class="blocktext">                        <p class="blocktitle">Time has passed</p><p class="blockdesc">Triggers after a specified amount of time</p>          </div></div></div><div class="blockelem create-flowy noselect"><input type="hidden" name="blockelemtype" class="blockelemtype" value="4"><div class="grabme"><img src="assets/grabme.svg"></div><div class="blockin">                    <div class="blockico"><span></span><img src="assets/error.svg"></div><div class="blocktext">                        <p class="blocktitle">Error prompt</p><p class="blockdesc">Triggers when a specified error happens</p>              </div></div></div>';
+            }
+            const canvas = document.getElementById("canvas");
+            if (canvas) {
+                this.flowyInstance = appn.flowy.create(canvas, this.drag.bind(this), this.release.bind(this), this.snapping.bind(this));
+            }
 
             this.addEventListenerMulti("click", this.disabledClick.bind(this), false, ".side");
-            document.getElementById("close").addEventListener("click", () => {
-                if (this.rightcard) {
-                    this.rightcard = false;
-                    document.getElementById("properties").classList.remove("expanded");
-                    setTimeout(() => {
-                        document.getElementById("propwrap").classList.remove("itson");
-                    }, 300);
-                    this.tempblock.classList.remove("selectedblock");
-                }
-            });
+            const close = document.getElementById("close");
+            if (close) {
+                close.addEventListener("click", () => {
+                    if (this.rightcard) {
+                        this.rightcard = false;
+                        document.getElementById("properties")?.classList.remove("expanded");
+                        setTimeout(() => {
+                            document.getElementById("propwrap")?.classList.remove("itson");
+                        }, 300);
+                        this.tempblock?.classList.remove("selectedblock");
+                    }
+                });
+            }
 
-            document.getElementById("removeblock").addEventListener("click", () => {
-                this.flowyInstance.deleteBlocks();
-            });
+            const removeblock = document.getElementById("removeblock");
+            if (removeblock) {
+                removeblock.addEventListener("click", () => {
+                    this.flowyInstance.deleteBlocks();
+                });
+            }
 
             document.addEventListener("mousedown", this.beginTouch.bind(this), false);
             document.addEventListener("mousemove", this.checkTouch.bind(this), false);
